refactor(git): Extract types to separate file
