+++ lib/manager/gradle-wrapper/artifacts-real.spec.ts
@@ -5,7 +5,7 @@ import * as httpMock from '../../../test/http-mock';
 import { git, partial } from '../../../test/util';
 import { GlobalConfig } from '../../config/global';
 import type { RepoGlobalConfig } from '../../config/types';
-import type { StatusResult } from '../../util/git';
+import type { StatusResult } from '../../util/git/types';
 import { ifSystemSupportsGradle } from '../gradle/deep/__testutil__/gradle';
 import type { UpdateArtifactsConfig } from '../types';
 import * as gradleWrapper from '.';
