+++ lib/workers/branch/execute-post-upgrade-commands.ts
@@ -5,7 +5,8 @@ import { addMeta, logger } from '../../logger';
 import type { ArtifactError } from '../../manager/types';
 import { exec } from '../../util/exec';
 import { readLocalFile, writeLocalFile } from '../../util/fs';
-import { File, getRepoStatus } from '../../util/git';
+import { getRepoStatus } from '../../util/git';
+import type { File } from '../../util/git/types';
 import { regEx } from '../../util/regex';
 import { sanitize } from '../../util/sanitize';
 import { compile } from '../../util/template';
